refactor/#25: application/port/in/command 패키지 생성 및 command 클래스들 위치 변경
