+++ src/main/java/com/yoger/productserviceorganization/product/application/port/in/DeductStockUseCase.java
@@ -1,5 +1,7 @@
 package com.yoger.productserviceorganization.product.application.port.in;
 
+import com.yoger.productserviceorganization.product.application.port.in.command.DeductStockCommandsFromOrderEvent;
+
 public interface DeductStockUseCase {
     void deductStockFromOrderCreated(DeductStockCommandsFromOrderEvent deductStockCommandsFromOrderEvent);
 }
