+++ src/main/java/com/yoger/productserviceorganization/product/application/port/in/IncreaseStockUseCase.java
@@ -1,5 +1,7 @@
 package com.yoger.productserviceorganization.product.application.port.in;
 
+import com.yoger.productserviceorganization.product.application.port.in.command.IncreaseStockCommand;
+
 public interface IncreaseStockUseCase {
     void increaseStockFromOrderCanceled(IncreaseStockCommand command);
 }
