+++ src/main/java/com/yoger/productserviceorganization/product/application/port/in/command/DeductStockCommandsFromOrderEvent.java
@@ -1,4 +1,4 @@
-package com.yoger.productserviceorganization.product.application.port.in;
+package com.yoger.productserviceorganization.product.application.port.in.command;
 
 import java.util.List;
